user.model.ts moved to the new structure
