+++ Servers/driver/autoDriver.driver.ts
@@ -11,7 +11,7 @@ import { createProjectRiskQuery } from "../utils/projectRisk.utils";
 import { createNewVendorQuery } from "../utils/vendor.utils";
 import { createNewVendorRiskQuery } from "../utils/vendorRisk.utils";
 import { createNewUserQuery } from "../utils/user.utils";
-import { User } from "../domain.layer/user/user.model";
+import { User } from "../domain.layer/models/user/user.model";
 import { Vendor } from "../models/vendor.model";
 
 import { createISOFrameworkQuery } from "../utils/iso42001.utils";
