+++ Servers/models/ISO-42001/annexCategoryISO.model.ts
@@ -5,7 +5,7 @@ import {
   Model,
   Table,
 } from "sequelize-typescript";
-import { UserModel } from "../../domain.layer/user/user.model";
+import { UserModel } from "../../domain.layer/models/user/user.model";
 import { ProjectFrameworksModel } from "../projectFrameworks.model";
 import { AnnexCategoryStructISOModel } from "./annexCategoryStructISO.model";
 import { Status, STATUSES } from "../../types/status.type";
