+++ Servers/models/ISO-42001/subClauseISO.model.ts
@@ -5,7 +5,7 @@ import {
   Model,
   Table,
 } from "sequelize-typescript";
-import { UserModel } from "../../domain.layer/user/user.model";
+import { UserModel } from "../../domain.layer/models/user/user.model";
 import { SubClauseStructISOModel } from "./subClauseStructISO.model";
 import { ProjectFrameworksModel } from "../projectFrameworks.model";
 import { Status, STATUSES } from "../../types/status.type";
