+++ Servers/models/file.model.ts
@@ -5,7 +5,7 @@ import {
   Model,
   Table,
 } from "sequelize-typescript";
-import { UserModel } from "../domain.layer/user/user.model";
+import { UserModel } from "../domain.layer/models/user/user.model";
 import { ProjectModel } from "./project.model";
 
 export interface File {
