+++ Servers/models/projectsMembers.model.ts
@@ -7,7 +7,7 @@ import {
   Table,
 } from "sequelize-typescript";
 import { ProjectModel } from "./project.model";
-import { UserModel } from "../domain.layer/user/user.model";
+import { UserModel } from "../domain.layer/models/user/user.model";
 
 export type ProjectsMembers = {
   user_id: number;
