+++ Servers/tsconfig.json
@@ -112,7 +112,7 @@
     "./structures/**/*.ts",
     "./types/**/*.d.ts",
     "types/status.type.ts",
-    "domain.layer/user/user.model.ts",
+    "domain.layer/models/user/user.model.ts",
   ],
   "exclude": [
     "node_modules",
